refactor(task): move TaskMapper and DTOs into task/ package
Continues the DDD migration: TaskMapper and its request/response DTOs move
from domain/ into task/dto/, no behavior change.
